# src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: louyi-admin 是一個輕量級的 Java 快速開發平臺，能快速開發項目並交付【接私活利器】
  version: 1.0.0
  title: 樓易健康科技管理系統
basePath: /louyi-admin
schemes:
  - http

#認證
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定義介面數據
paths:
...
  /sys/menu/info/{menuId}:
    get:
      tags:
        - 菜單管理
      summary: 獲取菜單信息
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜單ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回菜單信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              menu:
                description: 菜單信息
                $ref: '#/definitions/SysMenuEntity'
  /sys/menu/save:
    post:
      tags:
        - 菜單管理
      summary: 添加菜單
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜單對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
        - 菜單管理
      summary: 修改菜單
      produces:
        - application/json
      parameters:
        - name: body
          description: 菜單對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysMenuEntityEdit'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete/{menuId}:
    post:
      tags:
        - 菜單管理
      summary: 刪除菜單
      produces:
        - application/json
      parameters:
        - name: menuId
          description: 菜單ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/config/list:
    get:
      tags:
        - 參數管理
      summary: 參數列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序欄位
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: key
          description: 參數名
          in: query
          type: string
      responses:
        '200':
          description: 返回參數列表
          schema:
            $ref: '#/definitions/SysConfigEntityList'
  /sys/config/info/{id}:
    get:
      tags:
        - 參數管理
      summary: 獲取參數信息
      produces:
        - application/json
      parameters:
        - name: id
          description: 參數ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回參數信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              config:
                description: 返回參數信息
                $ref: '#/definitions/SysConfigEntity'
  /sys/config/save:
    post:
      tags:
        - 參數管理
      summary: 添加參數
      produces:
        - application/json
      parameters:
        - name: body
          description: 參數對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysConfigEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/update:
    post:
      tags:
        - 參數管理
      summary: 修改參數
      produces:
        - application/json
      parameters:
        - name: body
          description: 參數對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysConfigEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/config/delete:
    post:
      tags:
        - 參數管理
      summary: 刪除參數
      produces:
        - application/json
      parameters:
        - name: body
          description: 參數ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/oss/list:
    get:
      tags:
        - 文件服務
      summary: 文件列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序欄位
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
      responses:
        '200':
          description: 返迴文件列表
          schema:
            $ref: '#/definitions/SysOssEntityList'
...
  /sys/oss/saveConfig:
    post:
      tags:
        - 文件服務
      summary: 保存雲存儲配置信息
      produces:
        - application/json
      parameters:
        - name: body
          description: 參數對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysCloudStorageEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/oss/upload:
    post:
      tags:
        - 文件服務
      summary: 上傳文件
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: file
          description: 文件
          in: formData
          type: file
          required: true
      responses:
        '200':
          description: 返迴文件列表
          schema:
            $ref: '#/definitions/FileUpload'
  /sys/oss/delete:
    post:
      tags:
        - 文件服務
      summary: 刪除文件
      produces:
        - application/json
      parameters:
        - name: body
          description: 文件ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/list:
    get:
      tags:
        - 定時任務
      summary: 定時任務列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 頁碼
          in: query
          type: integer
          required: true
        - name: limit
          description: 每頁條數
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序欄位
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: beanName
          description: spring bean名稱
          in: query
          type: string
      responses:
        '200':
          description: 返回定時任務列表
          schema:
            $ref: '#/definitions/ScheduleJobEntityList'
  /sys/schedule/info/{jobId}:
    get:
      tags:
        - 定時任務
      summary: 獲取定時任務信息
      produces:
        - application/json
      parameters:
        - name: jobId
          description: 定時任務ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回定時任務信息
          schema:
            type: object
            properties:
              code:
                description: 狀態碼 0：成功  非0：失敗
                type: integer
                format: int32
              schedule:
                description: 定時任務信息
                $ref: '#/definitions/ScheduleJobEntity'
  /sys/schedule/save:
    post:
      tags:
        - 定時任務
      summary: 添加定時任務
      produces:
        - application/json
      parameters:
        - name: body
          description: 定時任務對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/ScheduleJobEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/update:
    post:
      tags:
        - 定時任務
      summary: 修改定時任務
      produces:
        - application/json
      parameters:
        - name: body
          description: 定時任務對象
          in: body
          type: string
          schema:
            $ref: '#/definitions/ScheduleJobEntity'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/delete:
    post:
      tags:
        - 定時任務
      summary: 刪除定時任務
      produces:
        - application/json
      parameters:
        - name: body
          description: 定時任務ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/run:
    post:
      tags:
        - 定時任務
      summary: 立即執行任務
      produces:
        - application/json
      parameters:
        - name: body
          description: 定時任務ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/pause:
    post:
      tags:
        - 定時任務
      summary: 暫停定時任務
      produces:
        - application/json
      parameters:
        - name: body
          description: 定時任務ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/schedule/resume:
    post:
      tags:
        - 定時任務
      summary: 恢復定時任務
      produces:
        - application/json
      parameters:
        - name: body
          description: 定時任務ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      msg:
        description: 失敗原因
        type: string
...
  SysMenuEntity:
    type: object
    properties:
      menuId:
        description: 菜單ID
        type: integer
        format: int64
      name:
        description: 菜單名稱
        type: string
      parentId:
        description: 父菜單ID，一級菜單為0
        type: integer
        format: int64
      parentName:
        description: 父菜單名稱
        type: string
      url:
        description: 菜單URL
        type: string
      perms:
        description: 授權標識
        type: string
      type:
        description: 類型     0：目錄   1：選單   2：按鈕
        type: integer
        format: int32
      icon:
        description: 菜單圖標
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展開   true：展開   false：不展開
        type: boolean
        format: int32
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜單ID
        type: integer
        format: int64
      name:
        description: 菜單名稱
        type: string
      parentId:
        description: 父菜單ID，一級菜單為0
        type: integer
        format: int64
      url:
        description: 菜單URL
        type: string
      perms:
        description: 授權標識
        type: string
      type:
        description: 類型     0：目錄   1：選單   2：按鈕
        type: integer
        format: int32
      icon:
        description: 菜單圖標
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
...
  SysConfigEntity:
    type: object
    properties:
      id:
        description: 參數ID
        type: integer
        format: int64
      key:
        description: 參數名
        type: string
      value:
        description: 參數值
        type: string
      remark:
        description: 備註
        type: string
  SysConfigEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysConfigEntity'
  SysOssEntity:
    type: object
    properties:
      id:
        description: ID
        type: integer
        format: int64
      url:
        description: URL地址
        type: string
      createTime:
        description: 創建時間
        type: string
        format: date-time
  SysOssEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysOssEntity'
  SysCloudStorageEntity:
    type: object
    properties:
      type:
        description: 類型 1：七牛  2：阿裡雲  3：騰訊雲
        type: integer
        format: int32
      qiniuDomain:
        description: 七牛綁定的功能變數名稱
        type: string
      qiniuPrefix:
        description: 七牛路徑前綴
        type: string
      qiniuAccessKey:
        description: 七牛ACCESS_KEY
        type: string
      qiniuSecretKey:
        description: 七牛SECRET_KEY
        type: string
      qiniuBucketName:
        description: 七牛存儲空間名
        type: string
      aliyunDomain:
        description: 阿裡雲綁定的功能變數名稱
        type: string
      aliyunPrefix:
        description: 阿裡雲路徑前綴
        type: string
      aliyunEndPoint:
        description: 阿裡雲EndPoint
        type: string
      aliyunAccessKeyId:
        description: 阿裡雲AccessKeyId
        type: string
      aliyunAccessKeySecret:
        description: 阿裡雲AccessKeySecret
        type: string
      aliyunBucketName:
        description: 阿裡雲BucketName
        type: string
      qcloudDomain:
        description: 騰訊雲綁定的功能變數名稱
        type: string
      qcloudPrefix:
        description: 騰訊雲路徑前綴
        type: string
      qcloudAppId:
        description: 騰訊雲AppId
        type: string
      qcloudSecretId:
        description: 騰訊雲SecretId
        type: string
      qcloudSecretKey:
        description: 騰訊雲SecretKey
        type: string
      qcloudBucketName:
        description: 騰訊雲BucketName
        type: string
      qcloudRegion:
        description: 騰訊雲COS所屬地區
        type: string
  FileUpload:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      url:
        description: 文件URL地址
        type: string
      msg:
        description: 失敗原因
        type: string
  ScheduleJobEntity:
    type: object
    properties:
      jobId:
        description: 任務ID
        type: integer
        format: int64
      beanName:
        description: spring bean名稱
        type: string
      methodName:
        description: 方法名
        type: string
      params:
        description: 參數
        type: string
      cronExpression:
        description: cron表達式
        type: string
      status:
        description: 任務狀態  0：正常  1：暫停
        type: integer
        format: int32
      remark:
        description: 備註
        type: string
      createTime:
        description: 創建時間
        type: string
        format: date-time
  ScheduleJobEntityList:
    type: object
    properties:
      code:
        description: 狀態碼 0：成功  非0：失敗
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 總記錄數
            type: integer
            format: int32
          pageSize:
            description: 每頁記錄數
            type: integer
            format: int32
          totalPage:
            description: 總頁數
            type: integer
            format: int32
          currPage:
            description: 當前頁數
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/ScheduleJobEntity'
...